pyproject.toml
[project]
name = "call-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "playsound>=1.3.0",
    "pydantic>=2.11.9",
    "pydantic-ai[anthropic]>=1.0.6",
    "pyobjc>=11.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.4",
    "devtools>=0.12.2",
    "ipython>=9.5.0",
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
extend-select = ["Q", "RUF100", "C90", "UP", "I"]
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
isort = { combine-as-imports = true }
mccabe = { max-complexity = 18 }

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportUnnecessaryTypeIgnoreComment = true
# reportIgnoreCommentWithoutRule = true
reportUnusedParameter = true
reportIncompatibleUnannotatedOverride = true
reportImplicitAbstractClass = true
include = ["call_alert"]
venv = ".venv"

[[tool.pyright.executionEnvironments]]
root = "call_alert/calendar_get.py"
reportMissingTypeStubs = false
reportUnknownVariableType = false
reportUnknownMemberType = false
